Addition of react components for navigation list and navigation item
The addition of the navigation list and item were added and attached to
the NavBar component. To change item names you must go into
NavItemList.js and change the itemName prop for each item. Styles have
yet to be added for item list & items.

Needed for users to navigate website.